requirements.txt
langchain==1.0.7
langgraph==1.0.3
langchain-google-genai==3.0.3
python-dotenv==1.2.1
requests==2.32.5
typing-extensions==4.15.0
streamlit>=1.28.0
